+++ apax/md/__init__.pyi
@@ -1,4 +1,4 @@
 from .ase_calc import ASECalculator
-from .nvt import run_md
+from .simulate import run_md
 
 __all__ = ["run_md", "ASECalculator"]
