+++ apax/nodes/md.py
@@ -9,7 +9,7 @@ import yaml
 import znh5md
 import zntrack.utils
 
-from apax.md.nvt import run_md
+from apax.apax.md.simulate import run_md
 
 from .model import Apax
 from .utils import check_duplicate_keys
